// src/app/core/swagger/model/userResource.ts
/**
 * CloudNote
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Role } from './role';

export interface UserResource {
    id: any;
    name: any;
    role: Role;
    created_at: any;
    updated_at: any;
}